shop/requirements.txt
Django==3.0.5
django-autoslug==1.9.7
django-bootstrap-form==3.4
django-crispy-forms==1.9.2
django-filter==2.3.0
django-map-widgets==0.3.0
django-phonenumber-field==4.0.0
liqpay-python==1.0
Pillow==7.1.1
pip==20.0.2
pytz==2019.3
requests==2.24.0
setuptools==46.1.3
sqlparse==0.3.1
urllib3==1.25.10
wheel==0.34.2